Update ns_maxemail.cpp
Remove redundant logic for writing that the email limit was reached.  The print will always print 1 or greater so there doesn't need to be a separate case for printing the same text with the number 1 filled in instead of NSEmailMax which has already been initialized.
Places the if (email.empty()) return 0; before initializing int count to save on allocating space and then deallocating when returning instead of counting.