hs_group: Prevent sending duplicate activated messages
The NickAlias passed to the Sync() function is:
* from CommandGroup: the NickAlias which has the existing vHost
* from OnSetVhost: the NickAlias that just had a vHost set
* from OnNickGroup: the NickAlias equal to the NickCore display

We don't need to re-set the vHost and call OnSetVhost for the given
NickAlias in any of the listed cases. Doing so can cause the
OnSetVhost hook in HostServCore to be called twice on the same
NickAlias. This also prevents sending an unnecessary message to
either the NickCore or source of GROUP, they already know they have
the vHost being used to sync to the rest.